+++ packages/connect-evm/src/index.ts
@@ -1,5 +1,5 @@
 export { getInfuraRpcUrls } from '@metamask/connect-multichain';
-export { createMetamaskConnectEVM, type MetamaskConnectEVM } from './connect';
+export { createEVMClient, type MetamaskConnectEVM } from './connect';
 export type { EIP1193Provider } from './provider';
 
 export type * from './types';