+++ packages/connect-multichain/src/index.native.ts
@@ -5,7 +5,7 @@ import { ModalFactory } from './ui/index.native';
 
 export * from './domain';
 
-export const createMetamaskConnect: CreateMultichainFN = async (options) => {
+export const createMultichainClient: CreateMultichainFN = async (options) => {
   const uiModules = await import('./ui/modals/rn');
   let storage: StoreClient;
   if (!options.storage) {