+++ playground/browser-playground/src/wagmi/metamask-connector.ts
@@ -17,7 +17,7 @@
 
 // @ts-nocheck
 import {
-  createMetamaskConnectEVM,
+  createEVMClient,
   type EIP1193Provider,
   type MetamaskConnectEVM,
 } from '@metamask/connect-evm';
@@ -48,14 +48,12 @@ import {
   withTimeout,
 } from 'viem';
 
-type CreateMetamaskConnectEVMParameters = Parameters<
-  typeof createMetamaskConnectEVM
->[0];
+type CreateEVMClientParameters = Parameters<typeof createEVMClient>[0];
 
 const DEFAULT_CHAIN_ID = 1;
 
 export type MetaMaskParameters = Partial<
-  Pick<CreateMetamaskConnectEVMParameters, 'dapp' | 'debug'>
+  Pick<CreateEVMClientParameters, 'dapp' | 'debug'>
 > &
   OneOf<
     | {
@@ -92,7 +90,7 @@ export function metaMask(parameters: MetaMaskParameters = {}) {
             ]),
           );
 
-          metamaskPromise = createMetamaskConnectEVM({
+          metamaskPromise = createEVMClient({
             dapp: parameters.dapp ?? { name: window.location.hostname },
             eventHandlers: {
               accountsChanged: connector.onAccountsChanged.bind(connector),