+++ playground/browser-playground/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Update to use renamed API functions: `createMetamaskConnect` → `createMultichainClient` and `createMetamaskConnectEVM` → `createEVMClient` ([#114](https://github.com/MetaMask/connect-monorepo/pull/114))
 - Remove `enable` flag from `analytics` config object for `createMetamaskConnect` constructor ([#92](https://github.com/MetaMask/connect-monorepo/pull/92))
 - Bump `@types/jest` from ^28.1.6 to ^29.5.14 ([#11](https://github.com/MetaMask/connect-monorepo/pull/11))
 - Bump `@types/node` from ^18.18 to ^22.9.0 ([#11](https://github.com/MetaMask/connect-monorepo/pull/11))